requirements.txt
# API & Server
fastapi
uvicorn[standard]
python-dotenv
pydantic

# LLM
langchain
langchain-groq
langchain-core

# Vector DB & RAG
chromadb
sentence-transformers
numpy

# arXiv ingestion
arxiv
pymupdf

# Document parsing
python-docx

# Frontend
streamlit
requests